backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
loguru
httpx
redis
qdrant-client
sentence-transformers
google-genai
passlib[bcrypt]
python-jose
sqlalchemy
bcrypt==4.0.1
